refactor: Move References to Module in H5Types
